requirements.txt
networkx
python-pptx